Move QR template tag to references app
