+++ references/templates/references/footer.html
@@ -1,4 +1,4 @@
-{% load rfid_tags %}
+{% load ref_tags %}
 <footer class="container mt-4 pt-4">
   <div class="row align-items-center">
     <div class="col-lg-3 text-center text-lg-start mb-3 mb-lg-0">
